LectureListItem VM 제거 - 3 (fin), 전면 UI화
